requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
djangorestframework==3.16.1
idna==3.11
mysqlclient==2.2.7
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
